Dialogue AutoTrigger implemented
Now each dialogue scriptable object has the option to be an "Auto Triggered dialogue" when such option is checked. The dialogue will trigger only once upon entering the collider and after this first occurrence, it will be only triggered if the player chooses to interact with the object holding the dialogue.